# product-api-service/handlers/products.go
package handlers

import (
	"net/http"
	"strconv"

	"github.com/arvindnama/golang-microservices/product-api-service/data"
	"github.com/hashicorp/go-hclog"

	"github.com/gorilla/mux"
)

type Products struct {
	l   hclog.Logger
	v   *data.Validation
	pDB *data.ProductsDB
}

type KeyProduct struct {
}

func NewProducts(
	l hclog.Logger,
	v *data.Validation,
	pDB *data.ProductsDB,
) *Products {
	return &Products{l, v, pDB}
}

//swagger:model GenericError
type GenericError struct {
	Message string `json:"message"`
}

// ValidationError is a collection error messages
//
//swagger:model ValidationError
type ValidationError struct {
	Messages []string `json:"messages"`
}

func getProductID(r *http.Request) int {
	vars := mux.Vars(r)

	id, err := strconv.Atoi(vars["id"])

	if err != nil {
		panic(err)
	}

	return id
}